feat: add DatabaseMetrics trait for generic DB in reth node
